# golang/temporal/apps/api/internal/http/v1/workflows_gen.go
// Package workflows provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package workflows

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Defines values for WorkflowRequestAction.
const (
	Start WorkflowRequestAction = "start"
	Stop  WorkflowRequestAction = "stop"
)

// WorkflowErrorResponse defines model for WorkflowErrorResponse.
type WorkflowErrorResponse struct {
	// Message Message
	Message string `json:"message"`
}

// WorkflowRequest defines model for WorkflowRequest.
type WorkflowRequest struct {
	// Action Action to be performed on the workflow
	Action WorkflowRequestAction `json:"action"`

	// Input Input data for the workflow
	Input *map[string]interface{} `json:"input,omitempty"`

	// Type Type of the workflow
	Type string `json:"type"`
}

// WorkflowRequestAction Action to be performed on the workflow
type WorkflowRequestAction string

// WorkflowResponse defines model for WorkflowResponse.
type WorkflowResponse struct {
	// WorkflowId ID of the workflow
	WorkflowId *string `json:"workflow_id,omitempty"`

	// WorkflowType Type of the workflow
	WorkflowType *string `json:"workflow_type,omitempty"`
}

// ManageWorkflowJSONRequestBody defines body for ManageWorkflow for application/json ContentType.
type ManageWorkflowJSONRequestBody = WorkflowRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Manage a workflow
	// (POST /v1/workflows)
	ManageWorkflow(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ManageWorkflow converts echo context to params.
func (w *ServerInterfaceWrapper) ManageWorkflow(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ManageWorkflow(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/v1/workflows", wrapper.ManageWorkflow)

}

type ManageWorkflowRequestObject struct {
	Body *ManageWorkflowJSONRequestBody
}

type ManageWorkflowResponseObject interface {
	VisitManageWorkflowResponse(w http.ResponseWriter) error
}

type ManageWorkflow200JSONResponse WorkflowResponse

func (response ManageWorkflow200JSONResponse) VisitManageWorkflowResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ManageWorkflow400JSONResponse WorkflowErrorResponse

func (response ManageWorkflow400JSONResponse) VisitManageWorkflowResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ManageWorkflow500JSONResponse WorkflowErrorResponse

func (response ManageWorkflow500JSONResponse) VisitManageWorkflowResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Manage a workflow
	// (POST /v1/workflows)
	ManageWorkflow(ctx context.Context, request ManageWorkflowRequestObject) (ManageWorkflowResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// ManageWorkflow operation middleware
func (sh *strictHandler) ManageWorkflow(ctx echo.Context) error {
	var request ManageWorkflowRequestObject

	var body ManageWorkflowJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ManageWorkflow(ctx.Request().Context(), request.(ManageWorkflowRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ManageWorkflow")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ManageWorkflowResponseObject); ok {
		return validResponse.VisitManageWorkflowResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}